requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.25.2          # pre-built wheel for py311
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.7.2
Pillow==10.0.1